Move frontend code to JS project - only kicker left in JVM project
